requirements.txt
# Brain 多模态感知基础设施 — 依赖清单
# conda create -n brain python=3.11 -y
# conda activate brain
# pip install -r requirements.txt

# ── 核心 ──
pydantic>=2.5
numpy>=1.24
scipy>=1.10

# ── 音频感知 ──
funasr_onnx>=0.4.0
# funasr>=1.0  # 完整版 ASR（可选，精度更高）

# ── 说话人分离 / 声纹 ──
pyannote.audio>=3.1
speechbrain>=1.0
sounddevice>=0.4

# ── PyTorch（GPU CUDA 12.6）──
# pip install torch torchaudio --index-url https://download.pytorch.org/whl/cu126
torch>=2.8
torchaudio>=2.8

# ── 视频处理 ──
# ffmpeg / ffprobe 需系统安装，非 pip 包

# ── 向量存储 ──
qdrant-client>=1.12
# pymilvus>=2.3  # Milvus（可选）

# ── MongoDB ──
pymongo>=4.6
mongomock>=4.1

# ── 测试 ──
pytest>=8.0
pytest-asyncio>=0.23
pytest-cov>=5.0
pytest-mock>=3.12

# ── 日志 / 工具 ──
pyyaml>=6.0
